requirements.txt
# 建筑物理 AI 世界模型 - 云端部署版本
# 生成命令: pip freeze > requirements.txt

# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
flask>=3.0.0
flask-cors>=4.0.0

# 深度学习
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.26.0
scikit-learn>=1.3.0

# 数据处理
pandas>=2.1.0
Pillow>=10.0.0
pydantic>=2.0.0

# PDF/文档解析
pymupdf>=1.23.0
pdfplumber>=0.10.0

# 可视化（可选，用于调试）
streamlit>=1.30.0
plotly>=5.18.0

# 其他工具
timm>=0.9.0
python-multipart>=0.0.6
